requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
